// packages/test/generatedCode/policies/claimsMappingPolicies/item/claimsMappingPolicyItemRequestBuilderGetRequestConfiguration.ts
import {ClaimsMappingPolicyItemRequestBuilderGetQueryParameters} from './claimsMappingPolicyItemRequestBuilderGetQueryParameters';
import {RequestOption} from '@microsoft/kiota-abstractions';

export interface ClaimsMappingPolicyItemRequestBuilderGetRequestConfiguration {
    /** Request headers */
    headers?: Record<string, string>;
    /** Request options */
    options?: RequestOption[];
    /** Request query parameters */
    queryParameters?: ClaimsMappingPolicyItemRequestBuilderGetQueryParameters;
}